[build-emacs] Upload to the high bandwidth server first.
This is because uploading to the main web server changes the web page to point to the
new version which then gets forwarded to the high bandwidth server. If we upload to
the high bandwidth server first then there's no "dead" time where a link is there but not
downloadable.