[combine-and-package,launch.rb] Segregate arch/version combo binaries. No more lipo.
Technically the previous version didn't actually support lipo, but had a
stub where it should be added.

Instead of going that route, this version just sticks everything in its own
directory. This makes the combiner really easy and doesn't lose much, except
maybe the cool factor of lipoing as much as possible.

launch.rb now has to pay attention to the architecture as well as the
version to figure out which emacs to run. It also now messes with the PATH
to add the architecture bin and libexec dirs since the main Emacs code
hardcodes them to Contents/MacOS/bin and Contents/MacOS/libexec,
respectively.