[launch.rb,combine-and-package] Use underscore instead of dot in version numbers to get around codesign bug
The 'codesign' app (from XCode 6.0) dies when it sees something like
'bin-i386-10.5' in the MacOS directory. I believe it thinks the directory is
a framework or some other kind of bundle and it dies when the contents don't
match up with what it expects. I've filed a bug with Apple (18340424) but in
the meantime, if we remove the dots from the name then codesign doesn't get
confused.