[ensure-brew] Fix stupid brew bootstrapping again.
Apparently brew doesn't download it's man formula until you actually try to
install or query things. Then it installs the homebrew/core tap. If we don't
do this manually then (again) it'll do this lazily when we try to query if a
package is installed and, because it outputs its progress to stdout, the
`brew info` json fails to parse.