Replace Ruby launcher with Rust launcher.
The Rust is longer but should be better for 2 reasons:

1. Starting with macOS 11 (or 10.15?), Apple warns the user when an app
   tries to access certain places on the filesystem (~/Documents, ~/Library,
   etc). With the Ruby launcher, the only way to allow this is by allowing
   /usr/bin/ruby, which is unintuitive and also allows more access than
   maybe you'd want. Having a native binary should make this process more
   clear and give permissions to only the intended app.

2. Apple has been threatening to remove the /usr/bin/ruby (and Perl and
   Python) binaries. That would mean we'd either have to bundle Ruby with
   Emacs (yuck), write it in shell (double yuck!), or make it a native
   binary.

There is another option: Remove the launcher entirely. Some day I hope to
get around to doing this. But it requires a lot of testing. Emacs can be
compiled against older Cocoa APIs but I think there are a few extra system
supplied libs (libxml2, for instance) that it links against. Those can
change across OS releases so to make it work everywhere we'd probably want
to bundle them (with the downside that they don't get security patches).