build-emacs-from-tar: Build with min os set to macOS 11 when building on macOS 12 with M1.
This is because I upgraded to macOS 12 without considering the ramifications
on the build system--people trying to launch the resulting emacs on M1 on
macOS 11 suddenly didn't have a workable binary.

Since there's no good macOS 11 virtualization story, I can never[1] go back
to macOS 11, so just set the minimum version on the cc command line. This
should probably be a command line option instead of naked logic in the code,
but it's 4am and I'm tired.