launch.rs: Fix environment dumping not actually dumping when launching from Finder.
A previous commit[1] broke the dumping in an unforseen way--since we stopped
calling `get_shell_environment()` when our parent wasn't "1", we didn't
actually get to the dumping code since the login shell we launched had us a
parent (duh). So when we launched from finder, we'd fork with a login shell
and the child would just launch emacs. Later, emacs would exit and the
launcher would try to parse _its_ output. 🤦

Fixes #117

[1] 3bb687b4cd5a3d55e663f34fbe2030c1374944f8 (launch.rs: Don't grab env vars from the login shell if we weren't launched from Finder)