combine-and-package: Notarize!
This changes codesigning so all exes are marked as using the hardened
runtime. The com.apple.security.cs.disable-library-validation entitlement is
gives so that Emacs can load shared libraries using the .so module loading
code.

Credentials must be stored in a keychain first, and are referenced by the
`-p` option (same as the xcrun notarytool option):

    xcrun notarytool store-credentials galvanix --apple-id <your-apple-id> --keychain <keychain-path> --team-id <team-id> --no-validate

--no-validate was essential because my AppleID doesn't have the Paid App
agreements in place. Those aren't necessary for notarizing but
store-credentials is for more than just notarizing, so it was being too
picky.