build-emacs-from-tar: copy_lib: Add support for @rpaths and frameworks
This looks like a rewrite even though it was a (painfully) slow metamorphosis
into something that worked and wasn't totally convoluted.

This was added because Nix on 10.14 x86_64 links against a non-system based
CoreFoundation.framework. On top of that it references it through an @rpath,
hence the support for both frameworks _and_ @rpath. In the end this adds an
extra 15M or so of unecessary libraries to the Emacs.app bundle on x86_64, which
is annoying but we'll get over it because it beats building all the dependencies
ourself.