launch.rs: Add path_append() which makes the main code cleaner
The old code was a little more succint, but I didn't like moving into a
String (lossily!) just to add the separators and then immediately moving it
back to an OsString.

path_append() takes Paths in to further remove boilerplate from the main
program and also so it can check if the path actually exists before adding
it.