dependencies.nix: Move x86_64 dependencies forward to update all the dependencies and backport tree-sitter 0.25.10
Unfortunately we can't move x86_64 all the way to current because they moved the minimum macOS version to 11.

The nixpkgs rev we had before had too old of a Rust compiler to compile
tree-sitter 0.25.10.

Luckily there exists a nixpkgs revision new enough that its Rust can compile
tree-sitter, but old enough that macOS 10.12 support has not been removed.

This dates the nixpkgs repo to around 2024-12. It appears that they got rid
of the 10.12 support somewhere at the end of that month, so the window is
fairly small. We could probably go a little more recent, but not by much.