dependencies.nix: Remove Nix dependency tar stuff. It's too limited. Replace with json file.
Having just the immediate dependencies isn't super helpful. It really
needs the entire set of transitive dependencies as source but those
are extremely hard to extract from Nix. Not only that it looks like
each build would produce about a gigabyte of dependencies since nix
includes nonsense like clang and the macOS sdks...

Instead, replace it with the list of dependency versions in a JSON
file.  The idea being that the web site could pick these up and
display them in some nice manner.